Moved PackageUpdateSet to where it's generated in nuget api folder
