+++ NuKeeper.Tests/RepositoryInspection/RepositoryScannerTests.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.IO;
-using System.Linq;
 using System.Reflection;
 using NuKeeper.RepositoryInspection;
 using NUnit.Framework;
