+++ NuKeeper/Engine/EngineReport.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using NuKeeper.Nuget.Api;
 using NuKeeper.RepositoryInspection;
 
 namespace NuKeeper.Engine
